requirements-dev.txt
black
coverage
pandas
pylint
pytest
pytest-env
typer-cli
werkzeug
xlrd==1.*
